@@ -9,7 +9,7 @@ describe ExportController do
   describe "GET 'export'" do
     it "returns http success" do
       get 'export'
-      response.should be_success
+      expect(response).to be_success
     end
   end
 end
