@@ -5,7 +5,7 @@ describe GrapesController do
   describe "GET 'search'" do
     it "returns http success" do
       get 'search'
-      response.should be_success
+      expect(response).to be_success
     end
   end
 end
