@@ -4,11 +4,11 @@ describe Bottle do
   describe "validations" do
     before do
       @bottle = Bottle.new
-      @bottle.should_not be_valid
+      expect(@bottle).not_to be_valid
     end
 
     it "must have a price" do
-      @bottle.errors.get(:price).should_not be_nil
+      expect(@bottle.errors.get(:price)).not_to be_nil
     end
   end
 
